requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.3
boto3==1.13.24
botocore==1.16.24
certifi==2018.8.24
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.1.1
defusedxml==0.5.0
Django==2.2
django-allauth==0.37.1
django-cors-headers==2.4.0
django-filter==2.3.0
django-rest-auth==0.9.3
django-seed==0.2.2
django-storages==1.9.1
django-templated-mail==1.1.1
django-test-curl==0.1.2
djangorestframework==3.8.2
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
docutils==0.15.2
Faker==4.1.3
gunicorn==19.7.1
idna==2.7
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==2.1.0
pep8==1.7.1
Pillow==7.1.2
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.0.0
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.1
urllib3==1.24.2
whitenoise==3.3.1
wrapt==1.12.1
